Do not power off after 180 seconds of inactivity
The original sample code powers off the system itself after 180 seconds of
inactivity.  After that, the user have to click a button to power on the
system again.  Since we do not have a button dedicated for that purpose, let's
just let the system run forever when it has some power supply.